First fully functional version of the script
This is the first fully functional version of the script, in which I have not been able to find any major bugs. The code comments are in Russian, as well as the text of the buttons. However, you can easily request any LLM to translate it into your preferred language for personal use.